Add files via upload
Finished adding the functionality I wanted. It's not as pretty as I'd have liked (using System.Drawing instead of SkiaSharp, because I was more familiar with it), but the knight's tour is completed. It's using Warnsdorff's rule, which guarantees finding a path for an 8x8 board, and the Knight's path itself is represented by blue squares.